fix!: harden auth, sessions, SSRF, stored XSS and fresh-install schema
Security:
- default RUST_ENV to production (fail closed); bacon/README set RUST_ENV=dev
- tibba-request: DNS resolver rejects internal targets (DNS rebinding SSRF)
- OAuth auto-merge only matches verified emails; email change clears
  verification; email verify tokens bound to {user_id, email} and one-shot
- stored XSS: data-driven response headers can no longer override
  content-type/disposition/CSP/set-cookie etc.; SVG served as attachment;
  file responses get CSP sandbox + nosniff applied last
- remove "first registered user becomes super admin"; bootstrap su via
  basic.super_admins at startup
- disabled accounts are rejected by password/2FA/JWT/OAuth/API key logins
- per-user session revocation (tibba-session revoke_user_sessions); applied
  on admin identity changes, user deletion and password reset
- permission soft delete cascades to role_permissions
- tibba-sql keeps sslmode and other PG connection params

Correctness:
- baseline migration for 16 legacy tables so fresh installs migrate
- update/delete by id return 404 when no row is affected
- sqlx/opendal errors mapped to proper status codes; LIKE escaping;
  saturating pagination offsets
- http_stat.list_by_created binds timestamps (query always failed before)
- connect_timeout applied as acquire_timeout; min <= max pool validation
- opendal database backend switched to postgres; config Debug redacts secrets
- tibba-error/util/config/crypto/runtime/cache refinements (status
  mapping, 413/415/422 params errors, device id validation, uuid v7,
  key rotation verification, processing limit 0 = unlimited)

BREAKING CHANGE: RUST_ENV defaults to production; first user is no longer
auto-promoted (configure basic.super_admins); several model APIs changed
(mark_email_verified, list_by_created, KeyGrip verification, AppState).

+++ tibba-cache/src/cache.rs
@@ -274,7 +274,7 @@ impl RedisCache {
 
     /// 获取指定键的剩余过期时间（秒）。
     /// 返回 -2 表示键不存在，-1 表示键无过期时间。
-    pub async fn ttl(&self, key: &str) -> Result<i32> {
+    pub async fn ttl(&self, key: &str) -> Result<i64> {
         let result = self
             .conn()
             .await?
